Refactor `downloadR` to fall back to old version
If the user passes an R version that is older than the current version, the CRAN url follows a different format. As a result, any time we want to download an older version, we need to first test the `currentURL` format, and if that fails, we try the `oldVersionURL` format before throwing a fatal error and exiting.